requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-telegram-bot>=21.0.1
python-dotenv>=1.0.0
aiohttp>=3.9.1
qrcode[pil]>=7.4.2
fpdf2>=2.7.6
Pillow>=10.2.0
